+++ packages/mip/test/specs/async-await.test.js
@@ -1,63 +1,64 @@
-import 'src/index'
+import main from 'src/index'
 
 describe('Async Await', function () {
-  it('should has not error', async function main() {
-    let arr = []
+  it('should has not error', function () {
+    return main()
+    // let arr = []
 
-    async function fn() {
-      try {
-        await new Promise((resolve, reject) => {
-          arr.push(function () {
-            console.log('a')
-            reject()
-          })
-        })
-      }
-      catch (e) {
-        console.log('c')
-      }
-    }
+    // async function fn() {
+    //   try {
+    //     await new Promise((resolve, reject) => {
+    //       arr.push(function () {
+    //         console.log('a')
+    //         reject()
+    //       })
+    //     })
+    //   }
+    //   catch (e) {
+    //     console.log('c')
+    //   }
+    // }
 
-    fn()
+    // fn()
 
-    await new Promise(resolve => {
-      arr.push(function () {
-        console.log('b')
-        resolve()
-      })
+    // await new Promise(resolve => {
+    //   arr.push(function () {
+    //     console.log('b')
+    //     resolve()
+    //   })
 
-      arr.forEach(item => item())
-    })
-    console.log('d')
+    //   arr.forEach(item => item())
+    // })
+    // console.log('d')
   })
 
-  it('should also has not error', function main() {
-    let arr = []
+  // it('should also has not error', function main() {
+  //   let arr = []
 
-    function fn() {
-      return new Promise((resolve, reject) => {
-        arr.push(function () {
-          console.log('a')
-          reject()
-        })
-      })
-      .catch(error => {
-        console.log('c')
-      })
-    }
+  //   function fn() {
+  //     return new Promise((resolve, reject) => {
+  //       arr.push(function () {
+  //         console.log('a')
+  //         reject()
+  //       })
+  //     })
+  //     .catch(error => {
+  //       console.log('c')
+  //     })
+  //   }
 
-    fn()
+  //   fn()
 
-    return new Promise(resolve => {
-      arr.push(function () {
-        console.log('b')
-        resolve()
-      })
+  //   return new Promise(resolve => {
+  //     arr.push(function () {
+  //       console.log('b')
+  //       resolve()
+  //     })
 
-      arr.forEach(item => item())
-    })
-    .then(() => {
-      console.log('d')
-    })
-  })
+  //     arr.forEach(item => item())
+  //   })
+  //   .then(() => {
+  //     console.log('d')
+  //   })
+  // })
 })
